Update comments and add alternative function name
Updated the comments for "connect_as_extension" and "create_new_application" to reflect the new return. I also added a function called "connect_as_standalone" which is identical to "create_new_application". In my opinion, the word "standalone" is relevant to make it a bit more consistent with the terms used by the ZOS API.